# swagger.partners.json
{
  "swagger": "2.0",
  "info": {
    "description": "Partners API for NFT marketplace integration.\n\nAuthorization header format: partners <token>\n\n**Rate Limits**\n\nA global limit of 25 req/s applies to all endpoints. Individual endpoints may have stricter limits:\n\n| Endpoint | Limit |\n|---|---|\n| GET /nfts/search | 15 req/s * |\n| GET /nfts/owned | 6 req/s |\n| GET /collections/floors | 9 req/s |\n| GET /collections/backdrops/floors | 2 req/s |\n| GET /collections/models/backgrounds/floors | 3 req/s |\n| GET /collections/attribute-floors | 5 req/s |\n| GET /collections/{id}/metrics | 8 req/s |\n| GET /collection-offers/{id}/top | 10 req/s * |\n| GET /collection-offers/{id}/all | 10 req/s * |\n\n\\* Can be increased on request. Contact support to adjust your rate limit.",
    "title": "Portals Partners API",
    "contact": {},
    "version": "1.0.0"
  },
  "host": "portal-market.com",
  "basePath": "/",
  "paths": {
    "/partners/collection-offers/": {
      "post": {
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "description": "Create an offer to buy NFTs from a specific collection. Funds will be frozen until offer is accepted or cancelled",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "collection-offers"
        ],
        "summary": "Create a new collection offer",
        "parameters": [
          {
            "description": "Collection offer details",
            "name": "request",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/httptrustedapps.CreateOfferRequest"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created",
            "schema": {
              "$ref": "#/definitions/model.CollectionOffer"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/exerr.HttpError"
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "schema": {
              "$ref": "#/definitions/exerr.HttpError"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/exerr.HttpError"
            }
          }
        }
      }
    },
    "/partners/collection-offers/cancel": {
      "post": {
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "description": "Cancel your own collection offer and return frozen funds",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "collection-offers"
        ],
        "summary": "Cancel a collection offer",
        "parameters": [
          {
            "description": "Collection offer ID",
            "name": "request",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/httptrustedapps.CancelOfferRequest"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/exerr.HttpError"
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/exerr.HttpError"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/exerr.HttpError"
            }
          }
        }
      }
    },
    "/partners/collection-offers/collections": {
      "get": {
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "description": "Returns list of collections that support collection offers",
        "produces": [
          "application/json"
        ],
        "tags": [
          "collection-offers"
        ],
        "summary": "Get collections available for offers",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/collectionoffer.FetchCollectionsOut"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/exerr.HttpError"
            }
          }
        }
      }
    },
...
    "/partners/collections/attribute-floors": {
      "get": {
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "description": "Returns floor prices for all unique model attributes across collections. Rate limit: 5 req/s.",
        "produces": [
          "application/json"
        ],
        "tags": [
          "collection-filters"
        ],
        "summary": "Get all unique attribute floor prices",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/collection.FetchAllUniqueAttributeFloorsOut"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/exerr.HttpError"
            }
          }
        }
      }
    },
    "/partners/collections/backdrops/floors": {
      "get": {
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "description": "Returns minimum listing prices grouped by backdrop type. Rate limit: 2 req/s.",
        "produces": [
          "application/json"
        ],
        "tags": [
          "collection-filters"
        ],
        "summary": "Get floor prices for common backgrounds",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/collection.FetchBackgroundFloorPricesOut"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/exerr.HttpError"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/exerr.HttpError"
            }
          }
        }
      }
    },
    "/partners/collections/floors": {
      "get": {
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "description": "Returns minimum listing prices for each collection. Rate limit: 9 req/s.",
        "produces": [
          "application/json"
        ],
        "tags": [
          "collection-filters"
        ],
        "summary": "Get floor prices for all collections",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/collection.FetchBackgroundFloorPricesOut"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/exerr.HttpError"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/exerr.HttpError"
            }
          }
        }
      }
    },
    "/partners/collections/models/backgrounds/floors": {
      "get": {
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "description": "Returns minimum listing prices for backdrops of specified models. Rate limit: 3 req/s.",
        "produces": [
          "application/json"
        ],
        "tags": [
          "collection-filters"
        ],
        "summary": "Get background floor prices for specific models",
        "parameters": [
          {
            "type": "string",
            "description": "comma-separated list of model names",
            "name": "models",
            "in": "query",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/collection.FetchModelsBackgroundsFloorPricesOut"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/exerr.HttpError"
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "schema": {
              "$ref": "#/definitions/exerr.HttpError"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/exerr.HttpError"
            }
          }
        }
      }
    },
    "/partners/collections/preview": {
      "get": {
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "description": "Returns all collections with preview images and basic info. No pagination",
        "produces": [
          "application/json"
        ],
        "tags": [
          "collections"
        ],
        "summary": "Get all collections with preview",
        "parameters": [
          {
            "type": "string",
            "description": "sort field (name, floor_price)",
            "name": "sort_by",
            "in": "query"
          },
          {
            "type": "string",
            "description": "sort order (asc|desc)",
            "name": "sort_order",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/collection.FetchAllCollectionsWithPreviewOut"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/exerr.HttpError"
            }
          }
        }
      }
    },
...
    "/partners/deposits": {
      "post": {
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "description": "Generate a unique deposit ID for TON deposits. Use this ID in transaction memo",
        "produces": [
          "application/json"
        ],
        "tags": [
          "deposit"
        ],
        "summary": "Generate deposit ID for user",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/userwallet.GenerateDepositIdOut"
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "schema": {
              "$ref": "#/definitions/exerr.HttpError"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/exerr.HttpError"
            }
          }
        }
      }
    },
...
    "/partners/market/config": {
      "get": {
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "description": "Returns market settings including commission rate, deposit wallet, and current TON/USDT rate",
        "produces": [
          "application/json"
        ],
        "tags": [
          "market"
        ],
        "summary": "Get market configuration",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/httptrustedapps.generalOut"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/exerr.HttpError"
            }
          }
        }
      }
    },
    "/partners/nfts": {
      "post": {
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "description": "Attempts to purchase multiple NFTs. Returns detailed results including successful purchases and failed attempts with reasons. Supports partial purchases - if some NFTs are invalid or unaffordable, others will still be purchased.",
        "produces": [
          "application/json"
        ],
        "tags": [
          "nfts"
        ],
        "summary": "Buy NFTs with partial purchase support",
        "parameters": [
          {
            "description": "body params",
            "name": "request",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/nft.BuyPartialIn"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/nft.BuyPartialOut"
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "schema": {
              "$ref": "#/definitions/exerr.HttpError"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/exerr.HttpError"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "collection.FetchBackgroundFloorPricesOut": {
      "type": "object",
      "required": [
        "floorPrices"
      ],
      "properties": {
        "floorPrices": {
          "$ref": "#/definitions/model.AttributeFloorPrices"
        }
      }
    },
...
    "collectionoffer.FetchCollectionsOut": {
      "type": "object",
      "properties": {
        "collections": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/model.Collection"
          }
        },
        "total_count": {
          "type": "integer"
        }
      }
    },
...
    "exerr.HttpError": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string"
        }
      }
    },
...
    "httptrustedapps.CancelOfferRequest": {
      "type": "object",
      "required": [
        "id"
      ],
      "properties": {
        "id": {
          "type": "string"
        }
      }
    },
    "httptrustedapps.CreateOfferRequest": {
      "type": "object",
      "required": [
        "amount",
        "collection_id",
        "max_nfts"
      ],
      "properties": {
        "amount": {
          "type": "string"
        },
        "collection_id": {
          "type": "string"
        },
        "expiration_days": {
          "type": "integer"
        },
        "max_nfts": {
          "type": "integer",
          "maximum": 50,
          "minimum": 1
        }
      }
    },
...
  },
  "securityDefinitions": {
    "ApiKeyAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  }
}